--- base.html
@@ -1,88 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Mountain Cruise Shuttle Bus System</title>
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
-
-    <!-- Optional JavaScript -->
-    <!-- jQuery first, then Popper.js, then Bootstrap JS -->
-    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.3/dist/umd/popper.min.js"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
-
-    <style>
-        body {
-            padding-top: 20px;
-        }
-        nav {
-            margin-bottom: 20px;
-        }
-    </style>
-</head>
-<body>
-    <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <a class="navbar-brand" href="#">Mountain Cruise</a>
-        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-            <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse" id="navbarNav">
-            <ul class="navbar-nav">
-                <li class="nav-item">
-                    <a class="nav-link" href="{% url 'index' %}">Home</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="{% url 'shuttle_list' %}">Shuttles</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="{% url 'ticket_list' %}">Tickets</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="{% url 'schedule_list' %}">Shuttle Schedules</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="{% url 'payment_details_list' %}">View Payment Details</a>
-                </li>
-            </ul>
-            <ul class="navbar-nav ml-auto">
-            {% if user.is_authenticated %}
-            <li class="nav-item">
-                <a class="nav-link" href="{% url 'logout' %}">Logout</a>
-            </li>
-            {% else %}
-            <li class="nav-item">
-                <a class="nav-link" href="{% url 'login' %}">Login</a>
-            </li>
-            {% endif %}
-        </ul>
-    </div>
-</nav>
-
-    <div class="container">
-        <main>
-            {% block content %}
-            {% endblock %}
-        </main>
-    </div>
-
-    <footer class="bg-light text-center text-lg-start">
-        <div class="container p-4">
-            <div class="row">
-                <div class="col-md-4">
-                    <p>&copy; 2023 Mountain Cruise Shuttle Bus System</p>
-                </div>
-                <div class="col-md-8">
-                    <ul class="list-inline text-right">
-                        <li class="list-inline-item"><a href="{% url 'privacy_policy' %}">Privacy Policy</a></li>
-                        <li class="list-inline-item"><a href="{% url 'terms_of_service' %}">Terms of Service</a></li>
-                        <li class="list-inline-item"><a href="{% url 'contact_us' %}">Contact Us</a></li>
-                    </ul>
-                </div>
-            </div>
-        </div>
-    </footer>
-</body>
-</html>
