--- index.html
@@ -1,6 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-    <h1>Welcome to Mountain Cruise Shuttle Bus System</h1>
-    <p>Explore our shuttles, view the schedule, or book a ticket.</p>
-{% endblock %}
