--- privacy_policy.html
@@ -1,33 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<h2>Privacy Policy</h2>
-
-<p>At Mountain Cruise Shuttle Bus System, we value your privacy. This policy explains how we use and protect
-    the personal data we collect.</p>
-
-<h3>1. Information We Collect</h3>
-<p>We may collect personal identification information including but not limited to your name, email address,
-    phone number, and payment details.</p>
-
-<h3>2. How We Use Your Information</h3>
-<p>Your information helps us to provide and improve our services. We may use the data for processing transactions,
-    for communication about bookings, and to enhance our platform based on user feedback.</p>
-
-<h3>3. Data Protection</h3>
-<p>We employ high-security measures to ensure your data is safe with us. Your personal and transaction data is stored
-    securely and is not shared with third parties, except where necessary to complete a transaction or as required by law.</p>
-
-<h3>4. Cookies</h3>
-<p>Our website uses cookies to enhance user experience. Cookies are small files that a site or its service provider
-    transfers to your device's hard drive through your web browser, enabling the site's systems to recognize your
-    browser and capture certain information.</p>
-
-<h3>5. Changes to Our Privacy Policy</h3>
-<p>Mountain Cruise Shuttle Bus System reserves the right to update this privacy policy at any time.
-    We will notify users of any changes by posting the new policy on this page.</p>
-
-<h3>6. Contact Us</h3>
-<p>If you have questions or concerns about our privacy practices, please contact us at support@mountaincruise.com.</p>
-
-{% endblock %}
