--- terms_of_service.html
@@ -1,31 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<h2>Terms of Service</h2>
-
-<p>Welcome to Mountain Cruise Shuttle Bus System. Please read these Terms of Service carefully before using our services.</p>
-
-<h3>1. Acceptance of Terms</h3>
-<p>By using our services, you agree to be bound by these terms. If you do not agree to these terms, please do not use
-    our services.</p>
-
-<h3>2. Registration</h3>
-<p>You may be required to register with our platform to access certain services. It is a condition of your use of
-    the services that all the information you provide is accurate, current, and complete.</p>
-
-<h3>3. Cancellations and Refunds</h3>
-<p>Customers can cancel or reschedule their shuttle up to 48 hours ahead of the scheduled departure.
-    Refunds will be processed according to our refund policy detailed in [Link to Refund Policy].</p>
-
-<h3>4. Limitation of Liability</h3>
-<p>Mountain Cruise Shuttle Bus System will not be liable for any indirect, incidental, or consequential damages
-    as a result of using our services.</p>
-
-<h3>5. Changes to Terms</h3>
-<p>We reserve the right to modify or replace these terms at any time. If a revision is material,
-    we will provide at least 30 days' notice prior to any new terms taking effect.</p>
-
-<h3>6. Contact</h3>
-<p>If you have any questions about these Terms, please contact us at support@mountaincruise.com.</p>
-
-{% endblock %}
