+++ ticketing/admin.py
@@ -1,10 +1,9 @@
 from django.contrib import admin
-from .models import Shuttle, Ticket, ShuttleSchedule, PaymentDetail
+from .models import Shuttle, Ticket, ShuttleSchedule, PaymentDetails
 # Register your models here.
 
 admin.site.register(Shuttle)
 admin.site.register(Ticket)
 admin.site.register(ShuttleSchedule)
-admin.site.register(PaymentDetail)
-admin.site.register(PaymentDetail)
+admin.site.register(PaymentDetails)
