+++ ticketing/templates/base.html
@@ -43,6 +43,9 @@
                 <li class="nav-item">
                     <a class="nav-link" href="{% url 'schedule_list' %}">Shuttle Schedules</a>
                 </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="{% url 'payment_details_list' %}">View Payment Details</a>
+                </li>
             </ul>
             <ul class="navbar-nav ml-auto">
             {% if user.is_authenticated %}
