+++ ticketing/templates/contact_us.html
@@ -1,0 +1,13 @@
+{% extends 'base.html' %}
+
+{% block content %}
+<h2>Contact Us</h2>
+<p>
+    For inquiries, suggestions, or complaints, you can reach us through the following means:
+</p>
+<ul>
+    <li>Email: support@mountaincruise.com</li>
+    <li>Phone: 402-456-7890</li>
+    <li>Address: 123 Mountain Rd, Cruisetown</li>
+</ul>
+{% endblock %}
