+++ ticketing/templates/payment_details_list.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Payment Details</title>
+</head>
+<body>
+    <h1>Payment Details</h1>
+    <p>Payment ID: {{ payment.id }}</p>
+    <p>Amount: ${{ payment.amount }}</p>
+    <p>Payment Date: {{ payment.payment_date }}</p>
+    <p>Payment Status: {{ payment.payment_status }}</p>
+    <p>Payment Method: {{ payment.payment_method }}</p>
+</body>
+</html>
