+++ ticketing/templates/shuttle_list.html
@@ -1,0 +1,10 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <h2>Shuttle List</h2>
+    <ul>
+    {% for shuttle in shuttle_list %}
+        <li>{{ shuttle.name }} - {{ shuttle.capacity }} seats</li>
+    {% endfor %}
+    </ul>
+{% endblock %}
