--- urls.py
@@ -1,13 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('', views.index, name='index'),
-    path('shuttles/', views.shuttle_list, name='shuttle_list'),
-    path('tickets/', views.ticket_list, name='ticket_list'),
-    path('schedule/', views.schedule_list, name='schedule_list'),
-    path('payments/', views.payment_details_list, name='payment_details_list'),
-    path('privacy_policy/', views.privacy_policy, name='privacy_policy'),
-    path('terms_of_service/', views.terms_of_service, name='terms_of_service'),
-    path('contact_us/', views.contact_us, name='contact_us'),
-]
